Rename file to full name
